Moved UIKit calls into a separate class
